+++ activesupport/test/core_ext/date_time_ext_test.rb
@@ -197,6 +197,17 @@ class DateTimeExtCalculationsTest < ActiveSupport::TestCase
     assert_equal DateTime.civil(2013, 10, 17, 20, 22, 19), DateTime.civil(2005, 2, 28, 15, 15, 10).advance(years: 7, months: 19, weeks: 2, days: 5, hours: 5, minutes: 7, seconds: 9)
   end
 
+  def test_advance_with_invalid_keys
+    d = DateTime.civil(2005, 2, 28, 15, 15, 10)
+    assert_raise(ArgumentError) { d.advance(year: 1) }
+    assert_raise(ArgumentError) { d.advance(month: 1) }
+    assert_raise(ArgumentError) { d.advance(week: 1) }
+    assert_raise(ArgumentError) { d.advance(day: 1) }
+    assert_raise(ArgumentError) { d.advance(hour: 1) }
+    assert_raise(ArgumentError) { d.advance(minute: 1) }
+    assert_raise(ArgumentError) { d.advance(second: 1) }
+  end
+
   def test_advance_partial_days
     assert_equal DateTime.civil(2012, 9, 29, 13, 15, 10),  DateTime.civil(2012, 9, 28, 1, 15, 10).advance(days: 1.5)
     assert_equal DateTime.civil(2012, 9, 28, 13, 15, 10),  DateTime.civil(2012, 9, 28, 1, 15, 10).advance(days: 0.5)